+++ views/usuario/detalles.php
@@ -10,6 +10,7 @@
 <body>
     <h1>Detalles</h1>
     <?php include '../views/components/menu.php'; ?>
+    <?php include '../views/components/login.php'; ?>
 
     <h2>Detalles del usuario</h2>
     <?= empty($GLOBALS['success']) ? "" : "<p style='color:#060'>" . $GLOBALS['success'] . "</p>" ?>
@@ -23,12 +24,18 @@
     <p><b>Apellido 1:</b> <?= $usuario->apellido1 ?></p>
     <p><b>Apellido 2:</b> <?= $usuario->apellido2 ?></p>
     <p><b>Privilegio:</b> <?= $usuario->privilegio ?></p>
-    <p><b>Administrador:</b> <?= $usuario->administrador ? 'Sí' : 'No'?></p>
+    <p><b>Administrador:</b> <?= $usuario->administrador ? 'Sí' : 'No' ?></p>
     <p><b>Correo electrónico:</b> <?= $usuario->email ?></p>
 
-    <a href="/usuario/edit/<?= $usuario->id ?>">Editar usuario</a>
-    <a href="/usuario/delete/<?= $usuario->id ?>">Borrar usuario</a>
-    <a href="/usuario/list">Lista de usuarios</a>
+    <?php
+    if (Login::isAdmin()) {
+    ?>
+        <a href="/usuario/edit/<?= $usuario->id ?>">Editar usuario</a>
+        <a href="/usuario/delete/<?= $usuario->id ?>">Borrar usuario</a>
+        <a href="/usuario/list">Lista de usuarios</a>
+    <?php
+    }
+    ?>
 </body>
 
 </html>
